Move caffe2 signal_handler to c10.
The signal_handler was under the caffe2 namespacee but was being used
by PyTorch as well.

I've fixed this my moving it to the c10 namespace where now both C2 and PyTorch
can use it.

The signal_handler interface in caffe2/utils/signal_handler.h is kept the same
for backward compatiblity for C2, but most of the commmon code is moved to c10.

Differential Revision: [D27946738](https://our.internmc.facebook.com/intern/diff/D27946738/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D27946738/)!

ghstack-source-id: 127192782
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56717